Move TrackerDeveloperToolsInterface to developer tools package
