+++ tracker/core/developer-tools/src/index.ts
@@ -3,6 +3,7 @@
  */
 
 import { makeGlobalDeveloperTools } from './makeGlobalDeveloperTools';
+import { TrackerDeveloperTools } from './TrackerDeveloperToolsInterface';
 
 export * from './validationRules/ContextValidationRuleConfig';
 export * from './validationRules/GlobalContextValidationRule';
@@ -12,8 +13,17 @@ export * from './ContextErrorMessages';
 export * from './ContextErrorType';
 export * from './ContextNames';
 export * from './makeGlobalDeveloperTools';
+export * from './TrackerDeveloperToolsInterface';
 export * from './types';
 
 const developerTools = makeGlobalDeveloperTools();
 
 export default developerTools;
+
+declare global {
+  var objectiv:
+    | undefined
+    | {
+        developerTools?: TrackerDeveloperTools;
+      };
+}
