+++ tracker/core/tracker/src/index.ts
@@ -2,8 +2,6 @@
  * Copyright 2021-2022 Objectiv B.V.
  */
 
-import { TrackerDeveloperTools } from './TrackerDeveloperToolsInterface';
-
 export * from './plugins/ApplicationContextPlugin';
 export * from './plugins/OpenTaxonomyValidationPlugin';
 
@@ -15,7 +13,6 @@ export * from './helpers';
 export * from './NoopConsoleImplementation';
 export * from './Tracker';
 export * from './TrackerConsole';
-export * from './TrackerDeveloperToolsInterface';
 export * from './TrackerElementLocations';
 export * from './TrackerEvent';
 export * from './TrackerPluginInterface';
@@ -34,11 +31,3 @@ export * from './TrackerTransportRetryAttempt';
 export * from './TrackerTransportSwitch';
 export * from './TrackerValidationRuleInterface';
 export * from './TrackerValidationLifecycleInterface';
-
-declare global {
-  var objectiv:
-    | undefined
-    | {
-        developerTools: TrackerDeveloperTools;
-      };
-}
